🗃Refactor: service,util function 삭제
