+++ src/cafe/index.ts
@@ -1,2 +1,1 @@
 export * from './store/Cafe';
-export * from './function/functions';
