+++ src/components/cafe/Modal/Modal.ts
@@ -4,7 +4,6 @@ import { Beverage, Order, Serving } from '@/domain';
 import { dispatchCustomEvent } from '@/common';
 import { EVENT } from '@/constant';
 import { OPTION_GROUP_NAMES } from '@/@types';
-import { getBeverageName } from '@/cafe';
 
 const CLASS_NAME_HIDDEN = 'hidden';
 
@@ -62,7 +61,7 @@ export class Modal extends Component {
 
           const serving = new Serving(
             servingOrder.getId(),
-            getBeverageName(servingOrder.getBeverageId()),
+            this.cafe.getBeverageName(servingOrder.getBeverageId()),
             servingOrder.getOrderTime(),
           );
