+++ src/constant/constant.ts
@@ -8,4 +8,5 @@ export const EVENT = {
   SERVING: 'SERVING',
   SERVED: 'SERVED',
   MODAL_OPEN: 'MODAL_OPEN',
+  MENU_BUTTON_CLICK: 'MENU_BUTTON_CLICK',
 };
